fix: edit style & rename directiry
