+++ src/components/Animes/AnimeItem/AnineItem.module.scss
