+++ src/components/Animes/AnimeItem/AnineItem.tsx
