+++ src/components/Animes/Anines.module.scss
@@ -10,6 +10,7 @@
 }
 
 .countries_title {
+  color: $color-default;
   padding-left: 4rem;
   padding-right: 4rem;
 }
