+++ src/components/Forms/Inputs/Input.module.scss
@@ -11,13 +11,14 @@
 .input_title {
   font-weight: bold;
   font-size: 1.17rem;
+  color: $color-btn-text;
 }
 
 .item_input {
   display: flex;
   flex-direction: column;
   margin-top: 0.1rem;
-  gap: 0.5rem;
+  gap: 0.1rem;
 
   &:first-child {
     margin-top: 2rem;
@@ -65,6 +66,9 @@
     padding-left: 40px;
     padding-top: 5px;
     vertical-align: middle;
+
+    font-size: 1rem;
+    color: $color-btn-text;
     &:after {
       transform: scale(0.5);
       content: '';
