+++ src/components/NavBar/NavBar.module.scss
@@ -1,7 +1,7 @@
 @import '/src/styles/variables';
 
 .navbar {
-  background: $background;
+  background: $backg-nav;
   height: 80px;
   display: flex;
   justify-content: flex-start;
