+++ src/index.scss
@@ -1,3 +1,5 @@
+@import '/src/styles/variables';
+
 * {
   box-sizing: border-box;
   margin: 0;
