+++ src/pages/Home.tsx
@@ -1,10 +1,11 @@
 import { Search } from '../components/Search/Search';
-import { Animes } from '../components/Countries/Animes';
+
 import { Pagination } from '../components/Pagination/Pagination';
 import { Spinner } from '../components/Spinner/Spinner';
 import { ErrorMessage } from '../components/Forms/ErrorMessage/ErrorMessage';
 import { useHome } from '../hooks/use-home';
 import styles from './Pages.module.scss';
+import { Animes } from '../components/Animes/Animes';
 
 export const Home = () => {
   const { animes, loading, error, search, handleSearchChange, handlerChangePage } = useHome();
