+++ src/styles/_colors.scss
@@ -7,3 +7,8 @@ $background: #141414;
 $firstColor: #014862;
 $secondColor: #0fa800;
 $white-50: #ffffff80;
+$backg-nav: #161b22;
+$backg-default: #0d1117;
+$color-input-bg: #21262d;
+$color-btn-text: #c9d1d9;
+$color-default: #e4e8ec;
